Previous calculation s + clear button
To get the previous problem I used the exact process as I did to get the numbers displayed but this time I added the operation used in it. I also created a new function that takes in a parameter and sets the previous value to current value which is nothing. To reset the numbers I just set the current value to an empty string an empty string then called it o the currentScreen.