Fixing Out of bounds error + adding decimal propety
The screen part of the calculator when you use to many numbers it go out of the screen to fix this I set a boundary of 10 individual numbers and checked to see if the value had exceeded that amount. If it did I used the slice method and add three dots to represent that there is more to the code. For the add decimal I used an EventListener and in it the addDecimal function is being called, which checks to see if the the current number doesn't already include a decimal so you don't get erros.